requirements.txt
torch==2.0.1
transformers==4.30.2
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
rdkit-pypi==2022.9.5
pillow==10.0.0
opencv-python==4.8.0.74
fastapi==0.100.0
uvicorn==0.23.2
pydantic==2.0.3
pyyaml==6.0
networkx==3.1
matplotlib==3.7.1
seaborn==0.12.2
torchvision==0.15.2
datasets==2.13.1
nltk==3.8.1
chembl-webresource-client==0.10.8